split runner test.rb into helper.rb
share common tests
